@@ -1,46 +1,45 @@
 {
   "name": "clean-electron-react",
-  "version": "0.1.0",
-  "main": "app/app.js",
-  "license": "MIT",
+  "productName": "Clean Electron React",
+  "version": "1.0.0",
+  "main": "./build/app.js",
   "scripts": {
     "watch": "webpack-dev-server",
     "dev": "electron .",
-    "build": "webpack && electron-builder"
+    "build": "webpack",
+    "package": "webpack && electron-builder"
   },
-  "dependencies": {
-    "electron-is-dev": "^1.2.0",
-    "react": "^17.0.1"
-  },
+  "license": "MIT",
+  "dependencies": {},
   "devDependencies": {
-    "@babel/core": "^7.12.3",
-    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
-    "@babel/preset-env": "^7.12.1",
-    "@babel/preset-react": "^7.12.1",
-    "@types/electron": "^1.6.10",
-    "@types/node": "^14.14.5",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
-    "babel-loader": "^8.1.0",
-    "electron": "^11.5.0",
-    "electron-builder": "^22.9.1",
-    "eslint": "^7.12.0",
-    "eslint-plugin-react": "^7.21.5",
-    "file-loader": "^6.1.1",
-    "html-webpack-plugin": "^4.5.0",
-    "prettier": "^2.1.2",
-    "react-dom": "^17.0.1",
-    "ts-loader": "^8.0.7",
-    "typescript": "^4.0.5",
-    "webpack": "^5.2.0",
-    "webpack-cli": "3.3.12",
-    "webpack-dev-server": "^3.11.0"
+    "@types/node": "^17.0.18",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
+    "electron": "^17.0.0",
+    "electron-builder": "^22.14.13",
+    "html-webpack-plugin": "^5.5.0",
+    "prettier": "^2.5.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "terser-webpack-plugin": "^5.3.1",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.5.5",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.68.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4"
   },
   "build": {
     "appId": "nonExistentID",
+    "directories": {
+      "buildResources": "assets"
+    },
     "files": [
       "build/**/*",
       "app/**/*"
-    ]
+    ],
+    "mac": {
+      "target": "dir"
+    }
   }
 }
